more components, fix public msgs
